Fix crossbuilds in VMs by setting ABUILD_TARGET
Without ABUILD_TARGET set rpmbuild gets called with --target
and as %_target_cpu and other stuff set to the hostarch.

We need to set ABUILD_TARGET for crossbuild quite early so
this information gets written to the VM by passing this environment
value via build/build.data to the guest.